+++ packages/ui/src/form/primitives/combobox/item.tsx
@@ -5,10 +5,6 @@ import { check, plus } from '@wordpress/icons';
 import { Icon } from '../../../icon';
 import itemPopupStyles from '../../../utils/css/item-popup.module.css';
 import resetStyles from '../../../utils/css/resets.module.css';
-import {
-	getItemPopupSizeClassName,
-	normalizeItemPopupSize,
-} from '../../../utils/item-popup-size';
 import type { ComboboxItemProps } from './types';
 
 export const Item = forwardRef< HTMLDivElement, ComboboxItemProps >(
@@ -22,14 +18,12 @@ export const Item = forwardRef< HTMLDivElement, ComboboxItemProps >(
 		},
 		ref
 	) {
-		const normalizedSize = normalizeItemPopupSize( size );
-
 		return (
 			<_Combobox.Item
 				className={ clsx(
 					resetStyles[ 'box-sizing' ],
 					itemPopupStyles.item,
-					getItemPopupSizeClassName( size, itemPopupStyles ),
+					size === 'small' && itemPopupStyles[ 'is-size-small' ],
 					className
 				) }
 				ref={ ref }
@@ -42,7 +36,7 @@ export const Item = forwardRef< HTMLDivElement, ComboboxItemProps >(
 						variant !== 'creatable' &&
 							itemPopupStyles[ 'item-indicator-icon' ]
 					) }
-					size={ normalizedSize === 'small' ? 20 : 24 }
+					size={ size === 'small' ? 20 : 24 }
 				/>
 				{ children }
 			</_Combobox.Item>